+++ src/gravity/mggravity.cpp
@@ -60,10 +60,9 @@ MGGravityDriver::MGGravityDriver(Mesh *pm, MGBoundaryFunc_t *MGBoundary,
 
   // Allocate multigrid objects
   RegionSize root_size=pmy_mesh_->mesh_size;
-  // std::int64_t nrbxi are unlikely to exceed std::int32_t limits when Multigrid is used
-  root_size.nx1 = static_cast<int>(pmy_mesh_->nrbx1);
-  root_size.nx2 = static_cast<int>(pmy_mesh_->nrbx2);
-  root_size.nx3 = static_cast<int>(pmy_mesh_->nrbx3);
+  root_size.nx1 = pmy_mesh_->nrbx1;
+  root_size.nx2 = pmy_mesh_->nrbx2;
+  root_size.nx3 = pmy_mesh_->nrbx3;
   LogicalLocation lroot;
   lroot.lx1=0, lroot.lx2=0, lroot.lx3=0, lroot.level=0;
   mgroot_= new MGGravity(this,lroot,-1,-1,root_size,MGBoundary,pmy_mesh_->mesh_bcs,true);
@@ -72,10 +71,9 @@ MGGravityDriver::MGGravityDriver(Mesh *pm, MGBoundaryFunc_t *MGBoundary,
   int nbs=nslist_[Globals::my_rank];
   int nbe=nbs+nblist_[Globals::my_rank]-1;
   RegionSize block_size;
-  // nrbx* are std::int64_t
-  block_size.nx1 = static_cast<int>(pmy_mesh_->mesh_size.nx1/pmy_mesh_->nrbx1);
-  block_size.nx2 = static_cast<int>(pmy_mesh_->mesh_size.nx2/pmy_mesh_->nrbx2);
-  block_size.nx3 = static_cast<int>(pmy_mesh_->mesh_size.nx3/pmy_mesh_->nrbx3);
+  block_size.nx1 = pmy_mesh_->mesh_size.nx1/pmy_mesh_->nrbx1;
+  block_size.nx2 = pmy_mesh_->mesh_size.nx2/pmy_mesh_->nrbx2;
+  block_size.nx3 = pmy_mesh_->mesh_size.nx3/pmy_mesh_->nrbx3;
   for (int i=nbs; i<=nbe; i++) {
     enum BoundaryFlag block_bcs[6];
     pmy_mesh_->SetBlockSizeAndBoundaries(pmy_mesh_->loclist[i], block_size, block_bcs);
